Move recorder impls out of `core`
