+++ tokio-trace-core/src/field.rs
@@ -64,7 +64,7 @@ use std::{
     error,
     fmt,
     hash::Hash,
-    io::{self, Write},
+    io,
 };
 
 pub type RecordResult = Result<(), Error>;
@@ -188,189 +188,6 @@ pub struct Error {
     // TODO: Should this carry a boxed error, or should it just discard the
     // cause and be opaque?
     inner: Box<dyn error::Error>,
-}
-
-pub struct DebugWriter<W> {
-    write: W,
-    comma_delimited: usize,
-    add_comma: bool,
-}
-
-impl<W> DebugWriter<W> {
-    pub fn into_inner(self) -> W {
-        self.write
-    }
-}
-
-impl<'a> DebugWriter<&'a mut fmt::Write> {
-    pub fn new_fmt<W: fmt::Write>(write: &'a mut W) -> Self {
-        Self {
-            write,
-            comma_delimited: 0,
-            add_comma: false,
-        }
-    }
-
-    fn fmt_maybe_comma(&mut self) -> fmt::Result {
-        if self.comma_delimited > 0 {
-            self.write.write_str(", ")?;
-        } else if self.add_comma {
-            self.comma_delimited += 1;
-            self.add_comma = false;
-        }
-        Ok(())
-    }
-
-    fn fmt_open(&mut self, args: fmt::Arguments) -> RecordResult {
-        self.fmt_maybe_comma()?;
-        self.write.write_fmt(args)?;
-        self.add_comma = true;
-        Ok(())
-    }
-
-    fn fmt_close(&mut self, args: fmt::Arguments) -> RecordResult {
-        self.write.write_fmt(args)?;
-        self.comma_delimited.saturating_sub(1);
-        Ok(())
-    }
-}
-
-impl<'a> DebugWriter<&'a mut io::Write> {
-    pub fn new_io<W: io::Write>(write: &'a mut W) -> Self {
-        Self {
-            write,
-            comma_delimited: 0,
-            add_comma: false,
-        }
-    }
-
-    fn io_maybe_comma(&mut self) -> io::Result<()> {
-        if self.comma_delimited > 0 {
-            self.write.write_all(b", ")?;
-        } else if self.add_comma {
-            self.comma_delimited += 1;
-            self.add_comma = false;
-        }
-        Ok(())
-    }
-
-    fn io_open(&mut self, args: fmt::Arguments) -> RecordResult {
-        self.io_maybe_comma()?;
-        self.write.write_fmt(args)?;
-        self.add_comma = true;
-        Ok(())
-    }
-
-    fn io_close(&mut self, args: fmt::Arguments) -> RecordResult {
-        self.write.write_fmt(args)?;
-        self.comma_delimited.saturating_sub(1);
-        Ok(())
-    }
-}
-
-impl<'a> Recorder for DebugWriter<&'a mut dyn io::Write> {
-    fn record_fmt(&mut self, args: fmt::Arguments) -> RecordResult {
-        self.io_maybe_comma()?;
-        self.write.write_fmt(args)?;
-        Ok(())
-    }
-
-    fn open_map(&mut self) -> RecordResult {
-        self.io_open(format_args!("{{"))
-    }
-
-    fn close_map(&mut self) -> RecordResult {
-        self.io_close(format_args!("}}"))
-    }
-
-    fn open_list(&mut self) -> RecordResult {
-        self.io_open(format_args!("["))
-    }
-
-    fn close_list(&mut self) -> RecordResult {
-        self.io_open(format_args!("]"))
-    }
-
-    fn open_struct(&mut self, name: &str) -> RecordResult {
-        self.io_open(format_args!("{} {{", name))
-    }
-
-    fn close_struct(&mut self) -> RecordResult {
-        self.io_close(format_args!("}}"))
-    }
-
-    fn open_tuple(&mut self) -> RecordResult {
-        self.io_open(format_args!("["))
-    }
-
-    fn close_tuple(&mut self) -> RecordResult {
-        self.io_open(format_args!("]"))
-    }
-
-    fn record_kv(&mut self, k: &dyn Value, v: &dyn Value) -> RecordResult {
-        self.io_maybe_comma()?;
-        k.record(self)?;
-        self.write.write_all(b": ")?;
-        v.record(self)?;
-        Ok(())
-    }
-
-    fn finish(mut self) -> RecordResult {
-        self.write.flush()?;
-        Ok(())
-    }
-}
-
-impl<'a> Recorder for DebugWriter<&'a mut dyn fmt::Write> {
-    fn record_fmt(&mut self, args: fmt::Arguments) -> RecordResult {
-        self.fmt_maybe_comma()?;
-        self.write.write_fmt(args)?;
-        Ok(())
-    }
-
-    fn open_map(&mut self) -> RecordResult {
-        self.fmt_open(format_args!("{{"))
-    }
-
-    fn close_map(&mut self) -> RecordResult {
-        self.fmt_close(format_args!("}}"))
-    }
-
-    fn open_list(&mut self) -> RecordResult {
-        self.fmt_open(format_args!("["))
-    }
-
-    fn close_list(&mut self) -> RecordResult {
-        self.fmt_open(format_args!("]"))
-    }
-
-    fn open_struct(&mut self, name: &str) -> RecordResult {
-        self.fmt_open(format_args!("{} {{", name))
-    }
-
-    fn close_struct(&mut self) -> RecordResult {
-        self.fmt_close(format_args!("}}"))
-    }
-
-    fn open_tuple(&mut self) -> RecordResult {
-        self.fmt_open(format_args!("["))
-    }
-
-    fn close_tuple(&mut self) -> RecordResult {
-        self.fmt_open(format_args!("]"))
-    }
-
-    fn record_kv(&mut self, k: &dyn Value, v: &dyn Value) -> RecordResult {
-        self.fmt_maybe_comma()?;
-        k.record(self)?;
-        self.write.write_str(": ")?;
-        v.record(self)?;
-        Ok(())
-    }
-
-    fn finish(mut self) -> RecordResult {
-        Ok(())
-    }
 }
 
 /// An opaque key allowing _O_(1) access to a field in a `Span` or `Event`'s
